Move DuckPage to separate file
