+++ duck_router/lib/duck_router.dart
@@ -6,4 +6,4 @@ export 'src/interceptor.dart';
 export 'src/location.dart';
 export 'src/duck_router.dart';
 export 'src/shell.dart';
-export 'src/pages/custom.dart';
+export 'src/pages/page.dart';
