+++ duck_router/lib/src/pages/pages.dart
@@ -1,3 +1,4 @@
 export 'cupertino.dart';
 export 'custom.dart';
 export 'material.dart';
+export 'page.dart';
